+++ lib/src/auth/authorization_server.dart
@@ -45,12 +45,12 @@ import 'auth.dart';
 ///             final router = new Router();
 ///             router
 ///               .route("/protected")
-///               .pipe(new Authorizer(authServer))
-///               .generate(() => new ProtectedResourceController());
+///               .link(() =>new Authorizer(authServer))
+///               .link(() => new ProtectedResourceController());
 ///
 ///             router
 ///               .route("/auth/token")
-///               .generate(() => new AuthController(authServer));
+///               .link(() => new AuthController(authServer));
 ///
 ///             return router;
 ///           }