+++ lib/src/http/router.dart
@@ -51,10 +51,10 @@ class Router extends Controller {
     _unmatchedController = listener;
   }
 
-  /// Adds a route to this instance.
+  /// Adds a route that [Controller]s can be linked to.
   ///
-  /// Requests that match [pattern] will be sent to the [Controller] returned by this method. Controllers that
-  /// should receive these requests should be attached to the returned [Controller] (via [pipe], [generate], or [listen]).
+  /// Routers allow for multiple linked controllers. A request that matches [pattern]
+  /// will be sent to the controller linked to this method's return value.
   ///
   /// The [pattern] must follow the rules of route patterns (see also http://aqueduct.io/docs/http/routing/).
   ///
@@ -88,6 +88,7 @@ class Router extends Controller {
 
   @override
   void prepare() {
+
     _rootRouteNode = new RouteNode(_routeControllers.expand((rh) => rh.specifications).toList());
 
     for (var c in _routeControllers) {
@@ -96,21 +97,13 @@ class Router extends Controller {
   }
 
   /// Routers override this method to throw an exception. Use [route] instead.
-  @override
-  Controller pipe(Controller n) {
-    throw new RouterException("Routers may not use pipe, use route instead.");
+  Controller link(Controller generatorFunction()) {
+    throw new StateError("Routers may not use generate, use route instead.");
   }
 
-  /// Routers override this method to throw an exception. Use [route] instead.
   @override
-  Controller generate(Controller generatorFunction()) {
-    throw new RouterException("Routers may not use generate, use route instead.");
-  }
-
-  /// Routers override this method to throw an exception. Use [route] instead.
-  @override
-  Controller listen(FutureOr<RequestOrResponse> handler(Request request)) {
-    throw new RouterException("Routers may not use listen, use route instead.");
+  Controller linkFunction(FutureOr<RequestOrResponse> handle(Request request)) {
+    throw new StateError("Routers may not use generate, use route instead.");
   }
 
   @override