+++ test/command/serve_test_project/lib/wildfire.dart
@@ -4,7 +4,7 @@ class WildfireChannel extends ApplicationChannel {
   @override
   Controller get entryPoint {
     final r = new Router();
-    r.route("/endpoint").listen((req) async => new Response.ok(null));
+    r.route("/endpoint").linkFunction((req) async => new Response.ok(null));
     return r;
   }
 }