+++ test/http/sink_test.dart
@@ -3,21 +3,6 @@ import 'package:test/test.dart';
 import 'package:aqueduct/aqueduct.dart';
 
 void main() {
-  test(
-      "Controller requiring instantion throws exception when instantiated early",
-      () async {
-    var app = new Application<TestChannel>();
-    try {
-      await app.start();
-      expect(true, false);
-    } on ApplicationStartupException catch (e) {
-      expect(
-          e.toString(),
-          contains(
-              "'FailingController' instances cannot be reused between requests. Rewrite as .generate(() => new FailingController())"));
-    }
-  });
-
   test("Find default ApplicationChannel", () {
     expect(ApplicationChannel.defaultType, equals(TestChannel));
   });
@@ -27,7 +12,7 @@ class TestChannel extends ApplicationChannel {
   @override
   Controller get entryPoint {
     final router = new Router();
-    router.route("/controller/[:id]").pipe(new FailingController());
+    router.route("/controller/[:id]").link(() =>new FailingController());
     return router;
   }
 }